Refactor: move some common functions to utils
